requirements.txt
Django==1.11.7
djangorestframework==3.7.1
pytz==2017.3
git+git://github.com/TKasekamp/OpyenXes@master#egg=opyenxes
pandas==0.21.0
numpy==1.13.3
scikit-learn==0.19.1
scipy==1.0.0
jsonfield==2.0.2
django-rq==1.0.0
django-cors-headers==2.0.0
